% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kamp_expectation.R
\name{get_kamp_expectation}
\alias{get_kamp_expectation}
\title{get_kamp_expectation}
\usage{
get_kamp_expectation(
  ppp_obj,
  rvec = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  markvar = "immune"
)
}
\arguments{
\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{rvec}{A vector of radii at which to calculate the KAMP expectation.}

\item{correction}{Type of border correction}

\item{markvar}{The variable used to mark the points in the point pattern object. Default is "immune".}
}
\value{
A tibble containing the sample_id, r values, K statistic, theoretical CSR, KAMP CSR, FUNDIFF
}
\description{
Calculates the KAMP expectation for a given point pattern object.
}
\details{
Title KAMP Expectation
}
