% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kamp_expectation.R
\name{get_kamp_expectation_mat}
\alias{get_kamp_expectation_mat}
\title{get_kamp_expectation_mat}
\usage{
get_kamp_expectation_mat(ppp_obj, rvec, markvar = "immune")
}
\arguments{
\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{rvec}{A vector of radii at which to calculate the KAMP expectation.}

\item{markvar}{The variable used to mark the points in the point pattern object. Default is "immune".}
}
\value{
A tibble containing the radius, expectation, and time taken to compute the expectation.
}
\description{
Wrapper function for mapping across multiple radii to calculate the KAMP expectation using matrices.
}
\details{
Title
}
