% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp.R
\name{get_kamp_expectation_mat_helper}
\alias{get_kamp_expectation_mat_helper}
\title{get_kamp_expectation_mat_helper}
\usage{
get_kamp_expectation_mat_helper(ppp_obj, rvalue, markvar = "immune")
}
\arguments{
\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{rvalue}{A single radius at which to calculate the KAMP expectation.}

\item{markvar}{The variable used to mark the points in the point pattern object. Default is "immune".}
}
\value{
A tibble containing the radius, expectation, and time taken to compute the expectation.
}
\description{
Helper function to calculate the KAMP expectation for a given point pattern object and radius.
}
\details{
Title KAMP Expectation Helper
}
