% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kamp_variance.R
\name{get_kamp_variance}
\alias{get_kamp_variance}
\title{get_kamp_variance}
\usage{
get_kamp_variance(ppp_obj, rvec, markvar = "immune")
}
\arguments{
\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{rvec}{A vector of radii at which to calculate the KAMP variance.}

\item{markvar}{The variable used to mark the points in the point pattern object. Default is "immune".}
}
\value{
A tibble containing the radius, K statistic, expectation, variance, Z statistic, p-value, and time taken to compute the variance.
}
\description{
Wrapper function for mapping across multiple radii, believe furrr is parallelized version but not sure i should do that
}
\details{
Title KAMP Variance
}
