% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_data.R
\name{get_sim_data}
\alias{get_sim_data}
\title{Title}
\usage{
get_sim_data(lambda_n, abundance, type = c("hom", "inhom"), clust = FALSE)
}
\arguments{
\item{lambda_n}{intensity for background cells}

\item{abundance}{intensity for marker positive cells}

\item{type}{defines the distribution of the point process- homogeneous, inhomogeneous, or clustered}

\item{clust}{should an image be simulated with or without holes}
}
\value{
A point pattern object of class "ppp" from the spatstat package where the two cell types are background and immune.
}
\description{
Title
}
