% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp_biv.R
\name{kamp_expectation_biv_mat_helper}
\alias{kamp_expectation_biv_mat_helper}
\title{kamp_expectation_biv_mat_helper}
\usage{
kamp_expectation_biv_mat_helper(
  ppp_obj,
  rvalue,
  correction = "trans",
  markvar1 = "immune1",
  markvar2 = "immune2"
)
}
\arguments{
\item{ppp_obj}{A point pattern object "ppp" from the \code{spatstat} package.}

\item{rvalue}{A single radius}

\item{correction}{Type of edge correction. Defaults to translational.}

\item{markvar1}{Variable used to mark the points in the point pattern object for the first type. Default is "immune1".}

\item{markvar2}{Variable used to mark the points in the point pattern object for the second type. Default is "immune2".}
}
\value{
A single-row dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value}
\item{theo_csr}{The theoretical K under CSR}
\item{kamp_csr}{The adjusted CSR representing the KAMP permuted expectation.}
\item{kamp_fundiff}{The difference between observed K and KAMP CSR}
}
}
\description{
Helper function to calculate the KAMP expectation for bivariate point patterns.
}
\details{
Helper function for KAMP Expectation (Bivariate matrix-based implementation)
}
\keyword{internal}
