% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp.R
\name{kamp_expectation_mat_helper}
\alias{kamp_expectation_mat_helper}
\title{kamp_expectation_mat_helper}
\usage{
kamp_expectation_mat_helper(
  ppp_obj,
  rvalue,
  correction = "trans",
  markvar = "immune"
)
}
\arguments{
\item{ppp_obj}{A point pattern object "ppp" from the spatstat package.}

\item{rvalue}{A single radius}

\item{markvar}{The variable used to mark the points in the point pattern object - default is "immune".}
}
\value{
A dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value from \code{Kcross}}
\item{theo_csr}{The theoretical K under CSR from \code{Kcross}}
\item{kamp_csr}{The adjusted CSR representing the permuted expectation.}
\item{kamp_fundiff}{The difference between observed K and KAMP CSR}
}
}
\description{
Helper function to calculate the KAMP expectation for a given point pattern object and radius.
}
\details{
KAMP Expectation Helper
}
\keyword{internal}
