% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp_biv.R
\name{kamp_variance_biv_helper}
\alias{kamp_variance_biv_helper}
\title{kamp_variance_biv_helper}
\usage{
kamp_variance_biv_helper(
  ppp_obj,
  rval,
  correction = "trans",
  mark1 = "immune1",
  mark2 = "immune2"
)
}
\arguments{
\item{ppp_obj}{A point pattern object "ppp" from the \code{spatstat} package.}

\item{rval}{A single radius}

\item{correction}{Type of edge correction. Defaults to translational.}

\item{mark1}{Variable used to mark the points in the point pattern object for the first type. Default is "immune1".}

\item{mark2}{Variable used to mark the points in the point pattern object for the second type. Default is "immune2".}
}
\value{
A single-row dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value}
\item{theo_csr}{The theoretical K under CSR}
\item{kamp_csr}{The adjusted CSR representing the KAMP permuted expectation.}
\item{var}{Variance of K under the permutation null distribution}
\item{pval}{P-value, calculated using the formula: pnorm(-z)}
}
}
\description{
Helper function to calculate the KAMP variance for bivariate point patterns.
}
\details{
Helper function for bivariate KAMP Variance
}
\keyword{internal}
