% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ovarian_df.R
\docType{data}
\name{ovarian_df}
\alias{ovarian_df}
\title{Ovarian Cancer Example Dataframe}
\format{
A dataframe with x and y coordinates and marks indicating immune status.
\describe{
\item{cell_id}{The ID of the current cell in the current sample/image}
\item{sample_id}{The ID of the sample/image}
\item{x}{X coordinate of the current cell}
\item{y}{Y coordinate of the current cell}
\item{immune}{A factor indicating whether the cell is immune or background}
\item{phenotype}{A factor indicating the cell type (e.g., b cell, cytotoxic t cell, helper t cell, macrophage, tumor, other)}
}
}
\source{
Subset of \code{HumanOvarianCancerVP()} from the VectraPolarisData package.
}
\usage{
ovarian_df
}
\description{
A processed dataframe derived from the VectraPolaris ovarian cancer dataset.
It is a snapshot of 5 images from the \code{HumanOvarianCancerVP()} dataset in the \code{VectraPolarisData} package
}
\keyword{datasets}
