% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pp_data.R
\name{sim_pp_data}
\alias{sim_pp_data}
\title{Simulates spatial data}
\usage{
sim_pp_data(
  lambda_n,
  abundance,
  cell_type1 = "immune",
  cell_type2 = "background",
  distribution = "hom",
  clust = FALSE
)
}
\arguments{
\item{lambda_n}{Number of total cells in image}

\item{abundance}{Percentage intensity for marker positive cells}

\item{cell_type1}{Marker positive cell type (default is "immune")}

\item{cell_type2}{Marker negative cell type (default is "background")}

\item{distribution}{Determines whether the image is homogeneous ("hom") or inhomogeneous ("inhom")}

\item{clust}{Determines whether an image is simulated with or without clustering (TRUE/FALSE)}
}
\value{
A point pattern object of class "ppp" from the spatstat package where the two cell types are background and immune.
}
\description{
Simulates a spatial point pattern with two cell types: background and immune.
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE) &&
   requireNamespace("ggplot2", quietly = TRUE) &&
   requireNamespace("tibble", quietly = TRUE) &&
   requireNamespace("magrittr", quietly = TRUE)) {
  pp_obj <- sim_pp_data(lambda_n = 200, abundance = 0.3)
  magrittr::`\%>\%`(
  tibble::as_tibble(pp_obj),
  ggplot2::ggplot(ggplot2::aes(x, y, color = marks))
  ) +
  ggplot2::geom_point()
}

}
