% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp.R
\name{check_inputs}
\alias{check_inputs}
\title{Checks inputs for KAMP functions}
\usage{
check_inputs(
  df,
  rvals,
  univariate,
  correction,
  mark_var,
  mark1,
  mark2,
  variance,
  thin,
  p_thin,
  background,
  ...
)
}
\arguments{
\item{df}{A dataframe containing the point pattern data. Will be converted into a \code{ppp} object.}

\item{rvals}{Vector of radius values at which to compute the KAMP expectation.}

\item{univariate}{Logical indicating whether to compute univariate KAMP expectation. Defaults to TRUE.}

\item{mark_var}{Column name in \code{df} that contains the marks for the point pattern object.}

\item{mark1}{Value used to mark the points in the point pattern object.}

\item{mark2}{Value used to mark the points in the point pattern object for the second type (optional, only used if \code{univariate} is FALSE).}

\item{variance}{Logical indicating whether to compute the variance of KAMP (default is FALSE).}

\item{thin}{Logical indicating whether to thin the point pattern before computing KAMP (default is FALSE), called KAMP-lite.}

\item{p_thin}{Percentage that determines how much to thin}

\item{background}{Value used to define the background for the point pattern object.}
}
\value{
TRUE if all inputs are valid, otherwise throws an error with a descriptive message.
}
\description{
Checks inputs for KAMP functions
}
\keyword{internal}
