% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp.R
\name{check_inputs}
\alias{check_inputs}
\title{Title}
\usage{
check_inputs(
  ppp_obj,
  rvals,
  univariate,
  correction,
  marksvar1,
  marksvar2,
  variance,
  thin,
  p_thin,
  background,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{internal}
