% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kamp.R
\name{get_kamp_variance_helper}
\alias{get_kamp_variance_helper}
\title{Title}
\usage{
get_kamp_variance_helper(rvalue, ppp_obj, correction = "trans")
}
\arguments{
\item{rvalue}{A radius at which to calculate the KAMP variance.}

\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{correction}{Type of border correction}
}
\value{
A tibble containing the radius, K statistic, expectation, variance, Z statistic, p-value, and time taken to compute the variance.
}
\description{
Title
}
