% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kperm-package.R
\docType{package}
\name{kamp-package}
\alias{kamp}
\alias{kamp-package}
\title{kamp: KAMP: Permutation-Based Ripley's K Analysis for Spatial Point Patterns}
\description{
Provides functions to calculate KAMP, a more robust and scalable K-statistic for inhomogenous spatial point pattern data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dliao1.github.io/kperm/}
}

}
\author{
\strong{Maintainer}: Daisy Liao \email{daisy.liao@emory.edu} (\href{https://orcid.org/0009-0006-3449-475X}{ORCID})

}
\keyword{internal}
