% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamp.R
\name{kamp}
\alias{kamp}
\title{Title}
\usage{
kamp(
  ppp_obj,
  rvals,
  univariate = TRUE,
  marksvar1,
  marksvar2 = NULL,
  correction = "trans",
  variance = FALSE,
  thin = FALSE,
  p_thin = 0.5,
  background = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
