% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamp_expectation.R
\name{kamp_expectation}
\alias{kamp_expectation}
\title{KAMP univariate expectation}
\usage{
kamp_expectation(
  ppp_obj,
  rvals = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  mark1 = "immune"
)
}
\arguments{
\item{ppp_obj}{A point pattern object from the \code{spatstat.geom} package.}

\item{rvals}{Vector of radii at which to calculate the KAMP expectation. Defaults to c(0, 0.05, 0.075, 0.1, 0.15, 0.2).}

\item{correction}{Type of edge correction method to be used and passed to \code{Kcross} and \code{Kest}. Defaults to translational edge correction.}

\item{mark1}{Identifies subset of marked points. Defaults to immune.}
}
\value{
A dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value from \code{Kcross}}
\item{theo_csr}{The theoretical K under CSR from \code{Kcross}}
\item{kamp_csr}{The adjusted CSR representing the permuted expectation.}
\item{kamp}{The difference between observed K and KAMP CSR}
}
}
\description{
Computes the KAMP (K-function Adjusted for Marked Permutations) expectation
for a given spatial point pattern. This function calculates Ripley's K
using both the traditional Ripley's K method (based on \code{Kcross})
and the KAMP-adjusted CSR baseline (based on \code{Kest}).

The KAMP-adjusted CSR represents a more robust baseline for K (compared
to traditional CSR) that accounts for spatial clustering or inhomogeneity
in a point pattern compared to the traditional CSR assumption, while
avoiding the computational burden of permuting the point pattern.

Notes:

This function uses the \code{spatstat} package under the hood, which
automatically uses border correction when the number of
points in the point pattern is more than 3000.

See \code{?Kcross} and \code{?Kest} for more details on the K calculation methods.
}
\details{
Compute KAMP Expectation
}
