% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamp_expectation_biv.R
\name{kamp_expectation_biv_mat}
\alias{kamp_expectation_biv_mat}
\title{KAMP bivariate expectation (Matrix Implementation)}
\usage{
kamp_expectation_biv_mat(
  ppp_obj,
  rvals = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  mark1 = "immune1",
  mark2 = "immune2"
)
}
\arguments{
\item{ppp_obj}{A point pattern object from the \code{spatstat.geom} package.}

\item{correction}{Type of edge correction method. Defaults to translational.}

\item{rvec}{Vector of radii at which to calculate the KAMP expectation. Defaults to c(0, 0.05, 0.075, 0.1, 0.15, 0.2).}

\item{markvar1}{Variable used to mark the points in the point pattern object for the first type. Default is "immune1".}

\item{markvar2}{Variable used to mark the points in the point pattern object for the second type. Default is "immune2".}
}
\value{
A dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value}
\item{theo_csr}{The theoretical K under CSR}
\item{kamp_csr}{The adjusted CSR representing the KAMP permuted expectation.}
\item{kamp}{The difference between observed K and KAMP CSR}
}
}
\description{
Computes the KAMP (K-function Adjusted for Marked Permutations) expectation
for bivariate point patterns using a matrix-based approach. Note that this
is slower.

See \code{kamp_expectation_biv} for the \code{spatstat}-based implementation of the KAMP
bivariate expectation.
}
\details{
Bivariate KAMP Expectation (Matrix Implementation)
}
