% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamp_variance.R
\name{kamp_variance}
\alias{kamp_variance}
\title{kamp_variance}
\usage{
kamp_variance(
  ppp_obj,
  rvec = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  markvar = "immune",
  thin_pct = 0
)
}
\arguments{
\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{rvec}{A vector of radii at which to calculate the KAMP variance.}

\item{correction}{Type of border correction (can either be translational or border)}

\item{markvar}{The variable used to mark the points in the point pattern object. Default is "immune".}

\item{thin_pct}{Percentage that determines how much to thin the amount of points in the point pattern object.}
}
\value{
A dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value}
\item{theo_csr}{The theoretical K under CSR}
\item{kamp_csr}{The adjusted CSR representing the KAMP permuted expectation.}
\item{kamp_fundiff}{The difference between observed K and KAMP CSR}
\item{var}{Variance of K under the permutation null distribution}
\item{z}{Z statistic, calculated by normalizing K using the formula: (K - KAMP)/sqrt(var)}
\item{pval}{P-value, calculated using the formula: pnorm(-z)}
}
}
\description{
Computes the KAMP (K-function Adjusted for Marked Permutations) variance
for a given spatial point pattern. Also returns the KAMP expectation,
z-statistic, and p-value.

Note: this a matrix-based implementation of the KAMP variance that
does not use the \code{spatstat} package.
}
\details{
KAMP Variance
}
\examples{
if (requireNamespace("spatstat.geom", quietly = TRUE)) {
  # simulates a simple spatial point pattern with two types
  win <- spatstat.geom::owin(c(0, 1), c(0, 1))
  pp <- spatstat.random::rpoispp(lambda = 100, win = win)
  marks <- sample(c("immune", "background"), pp$n, replace = TRUE)
  marked_pp <- spatstat.geom::ppp(pp$x, pp$y, window = win, marks = factor(marks))

  # computes KAMP variance
  kamp_result <- kamp_variance(marked_pp, markvar = "immune")
  print(kamp_result)
}
}
