% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamp_variance_Rcpp.R
\name{kamp_variance_Rcpp}
\alias{kamp_variance_Rcpp}
\title{KAMP univariate variance - Rcpp Version}
\usage{
kamp_variance_Rcpp(
  ppp_obj,
  rvals = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  mark1 = "immune"
)
}
\arguments{
\item{ppp_obj}{A point pattern object of class "ppp" from the spatstat package.}

\item{rvals}{A vector of radii at which to calculate the KAMP variance.}

\item{correction}{Type of edge correction. Defaults to translational.}

\item{mark1}{The variable used to mark the points in the point pattern object. Default is "immune".}
}
\value{
A dataframe with the following columns:
\describe{
\item{r}{The radius at which K was calculated.}
\item{k}{The observed K value}
\item{theo_csr}{The theoretical K under CSR}
\item{kamp_csr}{The adjusted CSR representing the KAMP permuted expectation.}
\item{kamp}{The difference between observed K and KAMP CSR}
\item{var}{Variance of K under the permutation null distribution}
\item{pval}{P-value, calculated using the formula: pnorm(-z)}
}
}
\description{
Computes the KAMP (K-function Adjusted for Marked Permutations) variance
for a given spatial point pattern. Also returns the KAMP expectation,
z-statistic, and p-value. Calculations are performed across a vector of
radii in a single call.

Note: this a matrix-based implementation of the KAMP variance. It relies on
an Rcpp function \code{kamp_pair_sums} to efficiently compute the necessary
pairwise sums for the KAMP variance calculation, as well as the \code{spatstat}
package for handling point pattern distances and edge corrections.
}
\details{
KAMP Variance
}
