% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ovarian_df.R
\docType{data}
\name{ovarian_df}
\alias{ovarian_df}
\title{Ovarian Cancer Example Point Pattern}
\format{
A \code{ppp} object (from the \code{spatstat.geom} package) with marks indicating immune status.
\describe{
\item{x}{X}
\item{y}{Y}
\item{marks}{Cell type: "immune" or "background"}
}
}
\source{
Subset of \code{HumanOvarianCancerVP()} from the VectraPolarisData package.
}
\usage{
ovarian_df
}
\description{
A processed spatial point pattern object derived from the VectraPolaris ovarian cancer dataset.
This object contains immune and background cells within the tumor region of one image.
}
\keyword{datasets}
